增加触摸画板的RGB565支持，修复更换笔触后颜色变回黑色的bug

+++ 22-电容触摸屏—触摸画板/User/touch/palette.h
@@ -114,7 +114,7 @@ typedef enum
 /*»­Ë¢²ÎÊý*/
 typedef struct
 {
-  uint16_t color;
+  uint32_t color;
   
   SHAPE  shape;
